CRITICAL FIX: Correct KtxTexture2 structure field sizes
Fixed structure marshaling bug that caused crashes when reading
Key-Value Data from KTX2 files.

Root cause: Boolean fields were declared as 'byte' (1 byte each),
but libktx uses 'ktx_bool_t' which is typedef'd to 'uint32_t' (4 bytes).

This caused a 12-byte misalignment:
- Expected: 4 flags × 4 bytes = 16 bytes
- Actual: 4 flags × 1 byte = 4 bytes
- Result: All subsequent fields read from wrong memory offsets

Evidence of bug:
- kvDataHead was reading as 0x6F (111) instead of valid pointer
- Previous debug showed kvDataHead=111, kvData=152 (invalid addresses)

Fix:
Changed fields from 'byte' to 'uint':
- public uint isArray
- public uint isCubemap
- public uint isCompressed
- public uint generateMipmaps

This aligns with ktxTexture structure definition in ktx.h from KTX-Software.

Impact: Fixes crash on "Found kvDataHead pointer: 0x6F" after previous commit.

+++ TextureViewer/LibKtxNative.cs
@@ -241,11 +241,11 @@ internal static class LibKtxNative {
         public IntPtr vvtbl;
         public IntPtr _protected;
 
-        // Flags (4 bytes each in the structure, but actually booleans)
-        public byte isArray;
-        public byte isCubemap;
-        public byte isCompressed;
-        public byte generateMipmaps;
+        // Flags (ktx_bool_t = uint32_t in C, NOT byte!)
+        public uint isArray;
+        public uint isCubemap;
+        public uint isCompressed;
+        public uint generateMipmaps;
 
         // Dimensions
         public uint baseWidth;
